requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
scipy==1.11.4
